requirements.txt
opencv-python==4.10.0.84
mmdet==3.3.0
mmcv==2.1.0
mmpose==1.3.2
numpy==1.24.4
tqdm==4.66.4
pyyaml==6.0.1